Fix array handling.
Previously if you fetched an record from MongoDB that contained an
array, it would get translated incorrectly into a Lua table with a [0]
index (rather than starting at 1). This would cause problems when trying
to iterate over arrays fetched from MongoDB.

Furthermore, when this adapter was used to insert arrays, the record
actually inserted in MongoDB was a hash object instead of an array. This
adapter would parse that type of object back as an array in Lua, so
that's why the current tests around arrays were passing. So in order to
properly test this, we'll use the mongo command line client to insert
some structures directly, and then make sure that real MongoDB arrays
get read back out as Lua array-type tables.

This also fixes broken tests around the replicaset getprimary()
function, since those were failing since MongoDB was returning an
array, which wasn't being iterated over properly.

This change does mean that an explicit [0] index type of table in Lua
has to change behavior, since we can't store that as a real Mongo array
and also know that the index should start at 0 in Lua. So these type of
arrays get converted to normal arrays upon writes, and then will get
read as an array (where the index will start with Lua's normal [1]
index). I'm not sure that's the best option (the "cjson" library
converts these types of [0] index arrays to hash-like objects upon
serialization instead of arrays), but it seems like this is the simplest
way to address these edge-case arrays for now.

See:
https://github.com/bigplum/lua-resty-mongol/issues/29
https://github.com/ashun/lua-resty-mongol/commit/9c4f9aaa223eb63f2505c24a2df150b1975b9f2f