Fix invalid query in TEST 1 of t/objid.t.
This started consistently breaking after the additional tests added in
f1575798dc99b2dc4d59cb90804dfc7db5cd3121. This test query wasn't
actually performing a conditional query, since it was querying on a nil
value (r["id"] rather than r["_id"]). So it was just sheer luck that the
test somehow used to fetch the correct record and pass. But the
additional records inserted as part of the new tests managed to unearth
this issue with the test.